+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 * Use correct working directory when using the `castor execute` command.
 * Fix infinite loop when executing `castor` with `castor execute jolicode/castor@castor` 
+* Fix `--castor-file` option parsing
 
 ### Internal